+++ json_serializable/test/kitchen_sink/kitchen_sink.g.dart
@@ -137,9 +137,9 @@ Map<String, dynamic> _$KitchenSinkToJson(KitchenSink instance) =>
 JsonConverterTestClass _$JsonConverterTestClassFromJson(
         Map<String, dynamic> json) =>
     JsonConverterTestClass(
-      durationConverter.fromJson(json['duration'] as int?),
+      const DurationMillisecondConverter().fromJson(json['duration'] as int?),
       (json['durationList'] as List<dynamic>)
-          .map((e) => durationConverter.fromJson(e as int?))
+          .map((e) => const DurationMillisecondConverter().fromJson(e as int?))
           .toList(),
       const BigIntStringConverter().fromJson(json['bigInt'] as String),
       (json['bigIntMap'] as Map<String, dynamic>).map(
@@ -159,6 +159,7 @@ JsonConverterTestClass _$JsonConverterTestClassFromJson(
           .map((e) => TrivialNumberConverter.instance.fromJson(e as int?))
           .toSet(),
       const EpochDateTimeConverter().fromJson(json['dateTime'] as int?),
+      trivialStringConverter.fromJson(json['trivialString'] as String?),
       TrivialNumberConverter.instance
           .fromJson(json['nullableNumberSilly'] as int?),
       (json['nullableNumberSillySet'] as List<dynamic>)
@@ -169,9 +170,11 @@ JsonConverterTestClass _$JsonConverterTestClassFromJson(
 Map<String, dynamic> _$JsonConverterTestClassToJson(
         JsonConverterTestClass instance) =>
     <String, dynamic>{
-      'duration': durationConverter.toJson(instance.duration),
-      'durationList':
-          instance.durationList.map(durationConverter.toJson).toList(),
+      'duration':
+          const DurationMillisecondConverter().toJson(instance.duration),
+      'durationList': instance.durationList
+          .map(const DurationMillisecondConverter().toJson)
+          .toList(),
       'bigInt': const BigIntStringConverter().toJson(instance.bigInt),
       'bigIntMap': instance.bigIntMap
           .map((k, e) => MapEntry(k, const BigIntStringConverter().toJson(e))),
@@ -187,6 +190,7 @@ Map<String, dynamic> _$JsonConverterTestClassToJson(
           .map(TrivialNumberConverter.instance.toJson)
           .toList(),
       'dateTime': const EpochDateTimeConverter().toJson(instance.dateTime),
+      'trivialString': trivialStringConverter.toJson(instance.trivialString),
       'nullableNumberSilly': _$JsonConverterToJson<int?, TrivialNumber>(
           instance.nullableNumberSilly, TrivialNumberConverter.instance.toJson),
       'nullableNumberSillySet': instance.nullableNumberSillySet