+++ json_serializable/test/kitchen_sink/kitchen_sink.g_any_map__checked.g.dart
@@ -183,12 +183,13 @@ JsonConverterTestClass _$JsonConverterTestClassFromJson(Map json) =>
       json,
       ($checkedConvert) {
         final val = JsonConverterTestClass(
-          $checkedConvert(
-              'duration', (v) => durationConverter.fromJson(v as int?)),
+          $checkedConvert('duration',
+              (v) => const DurationMillisecondConverter().fromJson(v as int?)),
           $checkedConvert(
               'durationList',
               (v) => (v as List<dynamic>)
-                  .map((e) => durationConverter.fromJson(e as int?))
+                  .map((e) =>
+                      const DurationMillisecondConverter().fromJson(e as int?))
                   .toList()),
           $checkedConvert('bigInt',
               (v) => const BigIntStringConverter().fromJson(v as String)),
@@ -220,6 +221,8 @@ JsonConverterTestClass _$JsonConverterTestClassFromJson(Map json) =>
                   .toSet()),
           $checkedConvert('dateTime',
               (v) => const EpochDateTimeConverter().fromJson(v as int?)),
+          $checkedConvert('trivialString',
+              (v) => trivialStringConverter.fromJson(v as String?)),
           $checkedConvert('nullableNumberSilly',
               (v) => TrivialNumberConverter.instance.fromJson(v as int?)),
           $checkedConvert(
@@ -236,9 +239,11 @@ JsonConverterTestClass _$JsonConverterTestClassFromJson(Map json) =>
 Map<String, dynamic> _$JsonConverterTestClassToJson(
         JsonConverterTestClass instance) =>
     <String, dynamic>{
-      'duration': durationConverter.toJson(instance.duration),
-      'durationList':
-          instance.durationList.map(durationConverter.toJson).toList(),
+      'duration':
+          const DurationMillisecondConverter().toJson(instance.duration),
+      'durationList': instance.durationList
+          .map(const DurationMillisecondConverter().toJson)
+          .toList(),
       'bigInt': const BigIntStringConverter().toJson(instance.bigInt),
       'bigIntMap': instance.bigIntMap
           .map((k, e) => MapEntry(k, const BigIntStringConverter().toJson(e))),
@@ -254,6 +259,7 @@ Map<String, dynamic> _$JsonConverterTestClassToJson(
           .map(TrivialNumberConverter.instance.toJson)
           .toList(),
       'dateTime': const EpochDateTimeConverter().toJson(instance.dateTime),
+      'trivialString': trivialStringConverter.toJson(instance.trivialString),
       'nullableNumberSilly': _$JsonConverterToJson<int?, TrivialNumber>(
           instance.nullableNumberSilly, TrivialNumberConverter.instance.toJson),
       'nullableNumberSillySet': instance.nullableNumberSillySet