+++ json_serializable/test/kitchen_sink/kitchen_sink.g_exclude_null.dart
@@ -71,6 +71,7 @@ class _Factory implements k.KitchenSinkFactory<String, dynamic> {
         TrivialNumber(0),
         {},
         DateTime.fromMillisecondsSinceEpoch(0),
+        TrivialString(''),
         TrivialNumber(0),
         {},
       );
@@ -196,13 +197,14 @@ class KitchenSink implements k.KitchenSink {
   }
 }
 
-@JsonSerializable(
-  includeIfNull: false,
-)
+@JsonSerializable(includeIfNull: false, converters: [
+  // referencing a top-level field should work
+  durationConverter,
+  // referencing via a const constructor should work
+  BigIntStringConverter(),
+])
 // referencing a top-level field should work
-@durationConverter
-// referencing via a const constructor should work
-@BigIntStringConverter()
+@trivialStringConverter
 @TrivialNumberConverter.instance
 @EpochDateTimeConverter()
 class JsonConverterTestClass implements k.JsonConverterTestClass {
@@ -216,6 +218,7 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
     this.numberSilly,
     this.numberSillySet,
     this.dateTime,
+    this.trivialString,
     this.nullableNumberSilly,
     this.nullableNumberSillySet,
   );
@@ -238,6 +241,8 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
   Set<TrivialNumber> numberSillySet;
 
   DateTime? dateTime;
+
+  TrivialString? trivialString;
 
   TrivialNumber? nullableNumberSilly;
   Set<TrivialNumber?> nullableNumberSillySet;